app/requirements.txt
streamlit==0.70.0
Pillow==7.2.0
pickleshare==0.7.5
tensorflow==2.3.1
numpy==1.18.5
